Added control for adding journeys.

+++ Journeys.Commands/AddJourneyCommand.cs
@@ -10,15 +10,15 @@ namespace Journeys.Commands
     {
         public Guid JourneyId { get; private set; }
 
-        public int JourneyDistance { get; private set; }
+        public decimal JourneyDistance { get; private set; }
 
         public DateTime JourneyDateOfOccurence { get; private set; }
 
         public string PersonName { get; private set; }
 
-        public int LiftDistance { get; private set; }
+        public decimal LiftDistance { get; private set; }
 
-        public AddJourneyCommand(Guid journeyId, int journeyDistance, DateTime journeyDateOfOccurence, string personName, int liftDistance)
+        public AddJourneyCommand(Guid journeyId, decimal journeyDistance, DateTime journeyDateOfOccurence, string personName, decimal liftDistance)
         {
             JourneyId = journeyId;
             JourneyDistance = journeyDistance;
